+++ packages/orq-rc/src/models/operations/listannotationqueues.ts
@@ -183,7 +183,7 @@ export const ListAnnotationQueuesData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-08-25T14:44:11.279Z",
+    "2026-08-26T07:29:36.190Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {